docs(devops): record the three boundary cases the evidence states must hold
Each of the three was wrong in the first draft of this contract and was caught
in review of PR #5641, so each is now stated with the reason it is not the
other state: an empty run blocks rather than exiting clean, applicability
outranks tool absence, and an advisory gate still separates a finding from a
failure. Adds aggregate.no_outcomes to the reason-code list, which the code
exported from the start and the document never named.

Also carries a test-file split that belongs with the previous commit and
landed here because an amend was applied to the wrong HEAD; interactive
rebase is unavailable in this environment, so the message is corrected rather
than the history. Adding the applicability class took
test_workflow_checks.py to 517 lines, past the 500-line taste ceiling, and
the taste ratchet absorbed it as an error-severity violation against its
remaining slack (565 of a 566 baseline). Extracting the class to
test_workflow_applicability.py returns the count to 564 and needs no
`# taste-lint: ignore` escape, which is the remedy ci-scripts.md prefers over
spending the baseline.

Refs #5635
